new functions - to add common article, and to add am/an to a word
anm(word): Add "an" or "am" to a word depending on first letter of the word. Note this is not just for definite articles, this word is used for a variety of things!

gd_common_article(word): add the common article ("the" in English) to a word, in the nominal, possessive or prepositional case, singular or plural. Currently a lot of function arguments, might simplify in the future if I can. Haven't included the vocative as it involves slenderisation, which can't be automated as it's entirely made out of exceptions. At some point I might have to put the vocative into the app, but I might put it into the spreadsheets.