Major code restructure
Instead of each lesson being separate code, I've unified it so that one piece of code provides the structure of a lesson (providing a question, marking the answer, calculating scores and time taken), and only the sentence/question generation is a separate piece of code. This standardised structure will make it much faster and easier to create new lessons. I've also started some documentation.