Prepositional case revision
Decided to remove slenderisation (for now) from the prepositional case. See documentation for full details - but essentially, slenderisation for the prepositional case is falling out of use and is considered to be quite formal.

I've added lots of detail to the documentation, updated a few other grammatical sections, and updated the code to remove slenderisation. Also words "an t-s..." don't lenite (in writing) so I've added that as an exception.

I've also removed the give_get lesson code since it's been replaced with give_to and get_from.